+++ lib/src/wrapper.dart
@@ -32,8 +32,9 @@ class TaleApiWrapper {
     String applicationInfo,
     String applicationDescription,
   ) async {
-      return api.auth(headers: await getHeaders(), 
-        applicationName: applicationName, 
+    return api.auth(
+        headers: await getHeaders(),
+        applicationName: applicationName,
         applicationInfo: applicationInfo,
         applicationDescription: applicationDescription);
   }
@@ -49,18 +50,22 @@ class TaleApiWrapper {
 
   Future<GameInfo> gameInfo() async {
     return api.gameInfo(headers: await getHeaders());
-  } 
+  }
 
   Future<PendingOperation> help() async {
     return api.help(headers: await getHeaders());
-  } 
-
-  Future<PendingOperation> checkOperation(String pendingUrl) async {
-      return api.checkOperation(pendingUrl, headers: await getHeaders());
   }
 
-  Future<String> getCards() async {
+  Future<PendingOperation> checkOperation(String pendingUrl) async {
+    return api.checkOperation(pendingUrl, headers: await getHeaders());
+  }
+
+  Future<CardList> getCards() async {
     return api.getCards(headers: await getHeaders());
+  }
+
+  Future<ReceivedCardList> receiveNewCards() async {
+    return api.receiveNewCards(headers: await getHeaders());
   }
 }
 
@@ -70,11 +75,9 @@ class WrapperBuilder {
     String applicationId,
     String appVersion,
   ]) {
-    return TaleApiWrapper(
-        TaleApi(
-            apiUrl: apiUrl,
-            applicationId: applicationId,
-            appVersion: appVersion),
-        apiUrl);
+    final taleApi = TaleApi(
+        apiUrl: apiUrl, applicationId: applicationId, appVersion: appVersion);
+
+    return TaleApiWrapper(taleApi, apiUrl);
   }
 }