+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-## 1.0.1
+## 1.0.4
+
+- Fixes `get-cards` return type.
+
+## 1.0.3
+
+- Adds `get-cards` request to api and wrapper.
+
+## 1.0.2
 
 - Adds example. Makes `auth` params required. Readme updates.